+++ crates/idxdb-store/src/ts/accounts.ts
@@ -285,7 +285,7 @@ export async function getAccountAuthByPubKey(pubKey: string) {
 
 // INSERT FUNCTIONS
 
-export async function insertAccountCode(codeRoot: string, code: Uint8Array) {
+export async function upsertAccountCode(codeRoot: string, code: Uint8Array) {
   try {
     // Prepare the data object to insert
     const data = {
@@ -300,7 +300,7 @@ export async function insertAccountCode(codeRoot: string, code: Uint8Array) {
   }
 }
 
-export async function insertAccountStorage(storageSlots: JsStorageSlot[]) {
+export async function upsertAccountStorage(storageSlots: JsStorageSlot[]) {
   try {
     let processedSlots = storageSlots.map((slot) => {
       return {
@@ -317,7 +317,7 @@ export async function insertAccountStorage(storageSlots: JsStorageSlot[]) {
   }
 }
 
-export async function insertStorageMapEntries(entries: JsStorageMapEntry[]) {
+export async function upsertStorageMapEntries(entries: JsStorageMapEntry[]) {
   try {
     let processedEntries = entries.map((entry) => {
       return {
@@ -333,7 +333,7 @@ export async function insertStorageMapEntries(entries: JsStorageMapEntry[]) {
   }
 }
 
-export async function insertVaultAssets(assets: JsVaultAsset[]) {
+export async function upsertVaultAssets(assets: JsVaultAsset[]) {
   try {
     let processedAssets = assets.map((asset) => {
       return {
@@ -349,7 +349,7 @@ export async function insertVaultAssets(assets: JsVaultAsset[]) {
     logWebStoreError(error, `Error inserting assets`);
   }
 }
-export async function insertAccountRecord(
+export async function upsertAccountRecord(
   accountId: string,
   codeRoot: string,
   storageRoot: string,
@@ -372,7 +372,7 @@ export async function insertAccountRecord(
       locked: false,
     };
 
-    await accounts.add(data as IAccount);
+    await accounts.put(data as IAccount);
   } catch (error) {
     logWebStoreError(error, `Error inserting account: ${accountId}`);
   }
@@ -402,7 +402,7 @@ export async function upsertForeignAccountCode(
   codeRoot: string
 ) {
   try {
-    await insertAccountCode(codeRoot, code);
+    await upsertAccountCode(codeRoot, code);
 
     const data = {
       accountId,